Refactor map-field-panel.js so it only has one init method
Previously the map initialisation was mostly in a function called "setAddress",
but also had some in the `connect` method. We've split the Controller
initialisation from the map initialisation, so it now hangs together a little
better.

Additionally, we now use controller properties in most places, rather than
passing a bunch of arguments that are already accessible by the callee